Modularize base class from initialization process; fix tests;
