+++ lib/base.js
@@ -12,7 +12,6 @@
 var configs = require('./configs'),
     fs = require('fs'),
     moment = require('moment'),
-    github = require('github'),
     path = require('path'),
     updateNotifier = require('update-notifier');
 
@@ -21,15 +20,6 @@ exports.clone = function(o) {
 };
 
 // -- Utils --------------------------------------------------------------------
-
-var baseConfig = configs.getConfig();
-
-exports.github = new github({
-    debug: false,
-    host: baseConfig.api.host,
-    protocol: baseConfig.api.protocol,
-    version: baseConfig.api.version
-});
 
 exports.checkVersion = function(opt_callback) {
     var notifier,
